add a new API to report report feature states
When inspecting features, the current API reports only
if each feature is active.
Client applications, especially those consuming the output
of `ethtool -k [iface]` also want to know if a feature can
actually be enabled or not. Examples:
- https://github.com/jaypipes/ghw/blob/main/README.md#network
-https://github.com/jaypipes/ghw/blob/main/pkg/net/net_linux.go#L161

This information is already available but is not exposed to the API.
So, add a new API to export all the flags. We add a new API instead
of changing the existing one to avoid breaking the backward
compatibility.

Signed-off-by: Francesco Romani <fromani@redhat.com>